ti-tidl-osrt: Use python3-dir for directory setting
This sets the python dir based on the python being deployed so we do not
have to update this recipe every time the python version is updated.

While here, inline PY_DST_DIR and LIB_DST_DIR as having a var for these
only adds an unnecessary layer of indirection.

Signed-off-by: Andrew Davis <afd@ti.com>