ti-tidl-osrt: Use unzip instead of native pip
These wheel format packages are actually zip files, and if nothing
extra is needed from pip we can use unzip to extract these. This
is a simpler operation. Pip will stop allowing installation in
global directories with the version in Yocto master, so this will
help avoid that issue when moving to master.

Signed-off-by: Andrew Davis <afd@ti.com>