linux-ti-staging: Include patches based on SoC name
The patches listed in the rtos-mem-map.inc file are added based on the
SoC name but the directories provided are based on the board name. This
means for any board not specifically listed the build will fail.

Instead list name these by SoC. This also removes the need to have
symlinks for the board variants we support, remove those.

Signed-off-by: Andrew Davis <afd@ti.com>